Simplify HomeScreen body by extracting method
Refactored the body of HomeScreen to its own method, _body. This improves the readability and maintainability of the code. Also, unused packages 'events.dart' and 'models.dart' have been removed leading to cleaner code. The change should not affect the functionality of the app, as it was purely for code organization purposes.